bots/polymarket/requirements.txt
python-dotenv
py-clob-client
web3
eth-account